+++ CHANGELOG.md
@@ -9,13 +9,14 @@ No date
 27.09.2023
 
 - Bump mist-cli version to 2.0.
-- Bump swiftDialog version to 2.3.2 except for systems running macOS 11 which still get 2.2.1.
+- Bump swiftDialog version to 2.3.2 except for systems running macOS 11 which still get 2.2.1. Note that the installer package includes version 2.3.2. If running on Big Sur, this will be deleted and an internet connection is required to download version 2.2.1.
 - Added localisation for Brazilian Portuguese (#432, thanks to @hooleahn).
 - `--os` searches will search for the relevant version name rather than number, to avoid a bug in mist-cli that may result in download an inappropriate installer if the chosen major OS is not available. Note that this bug is fixed in mist-cli 2.0, but I'll leave the workaround in place for the time being.
 - Use icons in the GitHub repo instead of using the InstallAssistant icon which occasionally doesn't render.
 - Stop trying to use `seedutil` on 13.4 or newer when using the `--ffi` option as it doesn't work any more.
 - Moved up the log rotation so that we get all output of the current run, and made it less verbose.
 - Add macOS Sonoma catalog.
+- Invalid package is removed when using `--replace-invalid` or `--update` regardless of whether the `--pkg` option is selected or not.
 
 ## [30.2]